Update table.py fix bug with "TypeError: fore|back|style() takes no arguments"
I debugged the fore|back|style errors: If it's for known and definite rendering effects like text, you can set it in the form of fore.GREEN. However, for unknown and dynamic content, it needs to be set as fg(set). If this is not modified, there will be an error when printing (robot). No other anomalies have been observed.